Remove implicit casts
Cast parameters of htonl in uint32_t in IN6_IS_ADDR_GLOBAL and
IN6_IS_ADDR_ULA definitions.
Remove comparison with 0 in while statement of vfmatch,
http_SendMessage and http_MakeMessage.
(cherry picked from commit 804088d85902305e949d9f7002d06b1b97a360da)

+++ upnp/src/api/upnpapi.c
@@ -84,14 +84,14 @@
 
 #ifndef IN6_IS_ADDR_GLOBAL
 #define IN6_IS_ADDR_GLOBAL(a) \
-		((((__const uint32_t *) (a))[0] & htonl(0x70000000)) \
-		== htonl (0x20000000))
+		((((__const uint32_t *) (a))[0] & htonl((uint32_t)0x70000000)) \
+		== htonl ((uint32_t)0x20000000))
 #endif /* IS ADDR GLOBAL */
 
 #ifndef IN6_IS_ADDR_ULA
 #define IN6_IS_ADDR_ULA(a) \
-		((((__const uint32_t *) (a))[0] & htonl(0xfe000000)) \
-		== htonl (0xfc000000))
+		((((__const uint32_t *) (a))[0] & htonl((uint32_t)0xfe000000)) \
+		== htonl ((uint32_t)0xfc000000))
 #endif /* IS ADDR ULA */
 
 /*! This structure is for virtual directory callbacks */
